move new query API to operators.js
